lift(track-a): interface/net Unicorn-prove 5 (ported:true).
Prove render_ui_widgets_postgame, multiplayer_game_set_text_box_for_teams_noteams,
FUN_00082bd0, interface_draw_splitscreen_dividers, hud_messaging_get_objective
at Unicorn 100/0/0; naked→C for splitscreen dividers + hud objective.

(cherry picked from commit 39dd4ada5cfe9254d8f5a8cbfd4f24cac8892fa6)

+++ src/halo/cache/tags.c
@@ -623,7 +623,7 @@ void FUN_001ba8b0(char *base)
   extern char DAT_002b839c[];
   extern char DAT_002b8488[];
   int status;
-  status = FUN_001d00b9(*(void **)(base + 0x950), 0x1388, 1);
+  status = FUN_001d00b9((int)(uintptr_t)*(void **)(base + 0x950), 0x1388, 1);
   if ((*(unsigned int *)(base + 0x994) >> 8) & 1) {
     display_assert(DAT_002b84a8, DAT_002b839c, 0x5ca, 1);
     system_exit(-1);
@@ -651,7 +651,7 @@ void FUN_001ba930(void *header)
     display_assert(DAT_002b8538, DAT_002b839c, 0x5d6, 1);
     system_exit(-1);
   }
-  status = FUN_001d00b9(*(void **)((char *)header + 0x950), 0x1388, 1);
+  status = FUN_001d00b9((int)(uintptr_t)*(void **)((char *)header + 0x950), 0x1388, 1);
   flags = *(unsigned int *)((char *)header + 0x994);
   if ((flags & 0x400u) != 0) {
     display_assert(DAT_002b84f0, DAT_002b839c, 0x5da, 1);
